requirements.txt
gevent>=1.0.2,<2.0.0
gevent-websocket>=0.9.5,<1.0.0
leancloud>=2.0.0,<3.0.0
Werkzeug>=0.11.11,<1.0.0
Flask>=0.10.1,<1.0.0
Flask-Sockets>=0.1,<1.0
bottlenose==1.1.3
amazon-scraper==0.3.3
beautifulsoup4==4.6.0
bottlenose==1.1.3
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
lxml==3.8.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
urllib3==1.22
xmltodict==0.11.0
fake-useragent==0.1.7
flask-bootstrap==3.3.7.1
flask-wtf==0.14.2
flask-moment==0.5.1
flask-mongoengine==0.9.3
